requirements.txt
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.1
